requirements.txt
gunicorn
whitenoise
psycopg2-binary
python-decouple
django-cors-headers
djangorestframework
djangorestframework-simplejwt